+++ plugins/set_value/expose_operation.h
@@ -41,12 +41,15 @@ class Expose_operation : public Operation
     std::vector<std::pair<std::string, PC_tree_t>> m_data_to_expose;
 
 public:
-
+    /** Creates expose operation
+     * \param[in] ctx context of the operation
+     * \param[in] list_of_values yaml config tree of operation
+     */
     Expose_operation(PDI::Context& ctx, PC_tree_t list_of_values);
     
     void execute() override;
 };
 
-} // namespace <anonymous>
+} // namespace set_value
 
 #endif //SET_VALUE_EXPOSE_OPERATION_H_