+++ plugins/set_value/operation.h
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2020 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
+ * Copyright (C) 2020-2021 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -30,19 +30,29 @@
 namespace set_value {
 
 class Operation {
+    /// Context of the operation
     PDI::Context& m_ctx;
 
 protected:
+    /** Getter of context
+     * \return context of the operation
+     */
     PDI::Context& context();
 
 public:
+    /** Creates operation
+     * \param[in] ctx context of the operation
+     * \param[in] release_value_node yaml config tree of operation
+     */
     Operation(PDI::Context& ctx);
-
+    
+    /// Triggers an operation
     virtual void execute() = 0;
 
+    /// Destroys operation
     virtual ~Operation() = default;
 };
 
-} // namespace <anonymous>
+} // namespace set_value
 
 #endif //SET_VALUE_OPERATION_H_