+++ plugins/set_value/set_operation.cxx
@@ -65,4 +65,4 @@ void Set_operation::execute()
 	}
 }
 
-}  // namespace <anonymous>
+}  // namespace set_value