+++ plugins/set_value/tests/set_value_test_06.c
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2020-2021 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
+ * Copyright (C) 2021 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,38 +22,24 @@
  * THE SOFTWARE.
  ******************************************************************************/
 
-#include <spdlog/spdlog.h>
+#include <assert.h>
 
-#include <pdi/context.h>
-#include <pdi/expression.h>
-#include <pdi/ref_any.h>
+#include <paraconf.h>
+#include <pdi.h>
 
-#include "share_operation.h"
-
-namespace set_value {
-
-Share_operation::Share_operation(PDI::Context& ctx, PC_tree_t list_of_values):
-    Operation{ctx}
+int main(int argc, char* argv[])
 {
-	size_t list_size = PDI::len(list_of_values);
-	context().logger()->debug("Share operation count: {}", list_size);
-	for (int i = 0; i < list_size; i++) {
-		PC_tree_t value_element = PC_get(list_of_values, "[%d]", i);
-		std::string data_name {PDI::to_string(PC_get(value_element, "{0}"))};
-		context().logger()->trace("\t {}: {}", i, data_name);
-		m_data_to_share.emplace_back(std::move(data_name), PC_get(value_element, "<0>"));
+	PDI_init(PC_parse_path(argv[1]));
+	
+	int size = 4;
+	PDI_expose("array_size", &size, PDI_OUT);
+	
+	int* int_array;
+	PDI_access("array_metadata", (void**)&int_array, PDI_IN);
+	for (int i = 0; i < 4; i++) {
+		assert(int_array[i] == i+1);
 	}
+	PDI_release("array_metadata");
+	
+	PDI_finalize();
 }
-
-void Share_operation::execute()
-{
-    for (auto& data_to_share : m_data_to_share) {
-		PDI::Data_descriptor& data_desc = context().desc(data_to_share.first);
-		PDI::Ref value_ref {PDI::Expression{data_to_share.second}.to_ref(context(), *data_desc.default_type()->evaluate(context()))};
-		context().logger()->trace("Sharing {} with size {} B", data_to_share.first, value_ref.type().buffersize());
-		data_desc.share(value_ref, false, false);
-		m_data_to_release.emplace(data_to_share.first);
-	}
-}
-
-}  // namespace set_value