+++ plugins/set_value/trigger.cxx
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2020 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
+ * Copyright (C) 2020-2021 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -26,9 +26,11 @@
 
 #include <pdi/error.h>
 
+#include "event_operation.h"
 #include "expose_operation.h"
 #include "set_operation.h"
 #include "share_operation.h"
+#include "release_operation.h"
 
 #include "trigger.h"
 
@@ -45,18 +47,18 @@ Trigger::Trigger(PDI::Context& ctx, PC_tree_t operation_list_node):
 	for (int i = 0; i < operations_count; i++) {
 		PC_tree_t value_map = PC_get(operation_list_node, "[%d]", i);
 		std::string operation = PDI::to_string(PC_get(value_map, "{0}"));
-		PC_tree_t operation_values = PC_get(value_map, "<0>");
-
-		if (!PDI::is_list(operation_values)) {
-			throw PDI::Config_error{"Operation value must be a list of values"};
-		}
+		PC_tree_t operation_value = PC_get(value_map, "<0>");
 
 		if (operation == "set") {
-			m_operations.emplace_back(new Set_operation{m_ctx, operation_values});
+			m_operations.emplace_back(new Set_operation{m_ctx, operation_value});
 		} else if (operation == "share") {
-			m_operations.emplace_back(new Share_operation{m_ctx, operation_values});
+			m_operations.emplace_back(new Share_operation{m_ctx, operation_value});
 		} else if (operation == "expose") {
-			m_operations.emplace_back(new Expose_operation{m_ctx, operation_values});
+			m_operations.emplace_back(new Expose_operation{m_ctx, operation_value});
+		} else if (operation == "event") {
+			m_operations.emplace_back(new Event_operation{m_ctx, operation_value});
+		} else if (operation == "release") {
+			m_operations.emplace_back(new Release_operation{m_ctx, operation_value});
 		} else {
 			throw PDI::Config_error{"Unknown operation: {}", operation};
 		}
@@ -70,4 +72,4 @@ void Trigger::execute()
 	}
 }
 
-}  // namespace <anonymous>
+}  // namespace set_value