+++ plugins/set_value/trigger.h
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2020 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
+ * Copyright (C) 2020-2021 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -50,6 +50,6 @@ public:
     void execute();
 };
 
-} // namespace <anonymous>
+} // namespace set_value
 
 #endif //SET_VALUE_TRIGGER_H_